requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
httpx
beautifulsoup4
prometheus_fastapi_instrumentator